% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Presents the available parameter IDs and a description of them.}
\usage{
get_parameters()
}
\value{
A dataframe with the ID and description info about the available parameters.
}
\description{
get_parameters() will provide you the available parameters for usage of the functions
smhi_data and smhi_plot. get_parameters() gives you the ID and a description of each parameter, i.e. 
what the different parameters will collect for you.
}
\examples{
##Get the description and ID of the available parameters.
get_parameters()

}
\references{
\url{http://opendata.smhi.se/apidocs/metobs/index.html}
}

