% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station.R
\name{get_station}
\alias{get_station}
\title{Get a list of the stations currently available for use of smhi_data and smhi_plot}
\usage{
get_station()
}
\value{
A dataframe with the ID and some additional info about the available weather stations.
}
\description{
get_station will prodvide you with a list of weather stations with fresh data. Only stations which are 
updated continously will appear, in order to be able to collect the most recent data.
To use the smhi_data and smhi_plot functions you need the ID of the station, which is presented along
with the stations' name, height, latitude and longitude.
}
\details{
Uses json to collect the data that's up to date, from SMHI.
}
\examples{
##Get the data of the available weather stations
get_station()

}
\references{
\url{http://opendata.smhi.se/apidocs/metobs/index.html}
}

