% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smhi_data.R
\name{smhi_data}
\alias{smhi_data}
\title{Fetch data from the SMHI API}
\usage{
smhi_data(station, parameter)
}
\arguments{
\item{station}{an ID for a weather station managed by SMHI. 
Available stations can be found by typing get_station()}

\item{parameter}{an ID for the type of data to collect. There are four different 
types available, which can be found by typing get_parameters()}
}
\value{
A dataframe with with the required type of data and the dates, over the last
four months.
}
\description{
Fetches data from the SMHI API via XML. The functions cleans the data and returns 
it in a data.frame. The function can collect data from weather stations managed by SMHI,
with requirement that the station are supplying data continously. 
The returning data is fram the last four months.
}
\details{
The function uses XML to collect data from the SMHI API. The data is 
transformed in order to contain a proper output. The "Date" column is in lubridate-format
}
\examples{
##Collect the air temperature hourwise (parameter=1) in Arvidsjaur(station=159880) the latest four months
smhi_data(159880,1)

}
\references{
\url{http://opendata.smhi.se/apidocs/metobs/index.html}
}

