% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smhi_plot.R
\name{smhi_plot}
\alias{smhi_plot}
\title{Plot data from the SMHI API}
\usage{
smhi_plot(station, parameter)
}
\arguments{
\item{station}{an ID for a weather station managed by SMHI. 
Available stations can be found by typing get_station()}

\item{parameter}{an ID for the type of data to collect. There are four different 
types available, which can be found by typing get_parameters()}
}
\value{
Generates and draws a timeseries-plot
}
\description{
Plots the latest avalible data from a station of your choosing. The function returns 
a timeseries-plot generated by ggplot2. Data is collected from active stations form the SMHI API.
}
\details{
The function uses the smhi_data function together with the get_station and get_parameters
to generate the data for the plot. The plot function used is ggplot2 together with the grid and 
gridExtra-packages for generating a publication-ready graph.
}
\examples{
##Plot the air temperature hourwise (parameter=1) in Arvidsjaur(station=159880) the latest four months
smhi_data(159880,1)

}
\references{
\url{http://opendata.smhi.se/apidocs/metobs/index.html}
}

