% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smhiAPI.R
\docType{package}
\name{smhiAPI}
\alias{smhiAPI}
\alias{smhiAPI-package}
\title{Access the SMHI API via R}
\description{
This package uses XML and JSON to access the SMHI API.
The package includes four functions, which will help you 
fetch and plot your weather data.
}

